requirements.txt
playwright>=1.40
asyncpg>=0.29
python-dotenv
yt-dlp>=2024.10.0
streamlit>=1.31
pandas>=2.0